Uninstall-AppxPackage

+++ itt.ps1
@@ -5822,7 +5822,7 @@ $sync.database.Tweaks = '[
         "Name": "Microsoft.WindowsCamera"
       },
       {
-        "Name": "microsoft.windowscommunicationsapps"
+        "Name": "Microsoft.windowscommunicationsapps"
       },
       {
         "Name": "Microsoft.WindowsFeedbackHub"
@@ -5916,6 +5916,12 @@ $sync.database.Tweaks = '[
       },
       {
         "Name": "Microsoft.MicrosoftStickyNotes"
+      },
+      {
+        "Name": "TikTok.TikTok_8wekyb3d8bbwe"
+      },
+      {
+        "Name": "TikTok"
       },
       {
         "Name": "Microsoft.ZuneMusic"
@@ -10006,21 +10012,18 @@ function Invoke-ApplyTweaks {
                                 $Name
                             )
                                 try {
-                                    #powershell.exe -Command "Import-Module Appx; Get-AppxPackage -AllUsers -Name "$($Name)" | Remove-AppxPackage -ErrorAction Stop"
-                                    #Start-Process powershell.exe -ArgumentList "-Command `"Import-Module Appx; Get-AppxPackage -AllUsers -Name '$($Name)' | Remove-AppxPackage -ErrorAction Stop`"" -NoNewWindow  -Wait 
-
-
-                                    Get-AppxPackage "$($Name)" | Remove-AppxPackage -ErrorAction SilentlyContinue
-                                    Get-AppXProvisionedPackage -Online | where DisplayName -EQ "$($Name)" | Remove-AppxProvisionedPackage -Online
-                                    Get-AppxPackage -AllUsers "$($Name)" | ForEach-Object {Add-AppxPackage -DisableDevelopmentMode -Register "$($_.InstallLocation)\AppXManifest.xml"}
-
-                                    #Start-Process powershell.exe -ArgumentList "-Command `"Get-AppXProvisionedPackage -Online | where DisplayName -EQ $($Name) | Remove-AppxProvisionedPackage -Online`"" -NoNewWindow  -Wait 
-
-
-                                    Add-Log -Message "Trying to remove $($Name)" -Level "INFO"
+                                    $checkapp = Get-AppxPackage -Name $($Name) -ErrorAction SilentlyContinue
+                                    if ($null -ne $checkapp) {
+                                        Add-Log -Message "Trying to remove $($Name)" -Level "INFO"
+                                        Get-AppxPackage "$($Name)" | Remove-AppxPackage -ErrorAction SilentlyContinue
+                                        Get-AppXProvisionedPackage -Online | where DisplayName -EQ "$($Name)" | Remove-AppxProvisionedPackage -Online
+                                        Get-AppxPackage -AllUsers "$($Name)" | ForEach-Object {Add-AppxPackage -DisableDevelopmentMode -Register "$($_.InstallLocation)\AppXManifest.xml"}
+                                    } else {
+                                        Write-Host "  $($Name) is not installed." -ForegroundColor Yellow
+                                    }
                                 } 
                                 catch {
-                                    Write-Host "Failed to remove $($Name). $_" -ForegroundColor red
+                                    Write-Host "  Failed to remove $($Name). $_" -ForegroundColor red
                                 }
                         }
